Clean up key prototype
